+++ pyunfold/__init__.py
@@ -2,3 +2,4 @@
 from .__version__ import __version__
 
 from .unfold import iterative_unfold
+from .callbacks import Logger