+++ docs/changelog.rst
@@ -11,7 +11,7 @@ Version 0.2.dev0 (TBD)
 
 **New Features**:
 
--
+- Added ``Logger`` callback to write test statistic information to stdout. (See `PR #18 <https://github.com/jrbourbeau/pyunfold/pull/18>`_)
 
 **Changes**: